+++ application/views/User/Dashboard.php
@@ -69,7 +69,7 @@
 </div>
 
 <?php foreach($trackerData as $trackerDataTypeKey => $trackerDataType) { ?>
-<table class="tablesorter tablesorter-bootstrap table-striped tracker-table" data-list="<?=$trackerDataTypeKey?>" style="<?=($trackerDataTypeKey !== 'reading' ? 'display:none' : '')?>">
+<table class="tablesorter tablesorter-bootstrap table-striped tracker-table" data-list="<?=$trackerDataTypeKey?>" style="<?=($trackerDataTypeKey !== 'reading' ? 'display:none' : '')?>" data-unread="<?=$trackerDataType['unread_count']?>">
 	<thead>
 		<tr>
 			<th class="header read headerSortDown"></th>